+++ src/test/kotlin/dev/meanmail/parsing/ParsingTest.kt
@@ -18,6 +18,7 @@ class ParsingTest : ParsingTestCase(
     fun testLua() {
         doTest(true)
     }
+
     fun testNestedLocation() {
         doTest(true)
     }