+++ src/main/kotlin/dev/meanmail/NginxLanguage.kt
@@ -2,4 +2,6 @@ package dev.meanmail
 
 import com.intellij.lang.Language
 
-object NginxLanguage : Language("Nginx Configuration")
+object NginxLanguage : Language("Nginx Configuration") {
+    private fun readResolve(): Any = NginxLanguage
+}