+++ src/main/kotlin/dev/meanmail/directives/nginx/http/ngx_http_ssl_module.kt
@@ -484,3 +484,11 @@ val sslVerifyDepth = Directive(
     context = listOf(http, server),
     module = ngx_http_ssl_module
 )
+
+val sslCertificateCompression = ToggleDirective(
+    name = "ssl_certificate_compression",
+    description = "Enables TLSv1.3 certificate compression (RFC 8879). Disabled by default since 1.29.1",
+    enabled = false,
+    context = listOf(http, server),
+    module = ngx_http_ssl_module
+)